V1.4 fixes: - Fixed outOfMemory problem caused by the example photo.

+++ app/src/main/java/dinidiniz/eggsearcher/App.java
@@ -1,14 +1,11 @@
 package dinidiniz.eggsearcher;
 
 import android.app.Application;
+import android.content.pm.PackageManager;
 import android.graphics.Bitmap;
 import android.graphics.drawable.BitmapDrawable;
-import android.os.Environment;
 import android.support.v4.content.ContextCompat;
-import android.util.Log;
 
-import java.io.BufferedOutputStream;
-import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
@@ -33,12 +30,14 @@ public class App extends Application {
 
         File storageDir = new File(Consts.getImagePath());
         File file = new File(storageDir, "example.jpg");
-        if (!file.exists()) {
-            storageDir.mkdirs();
-            BitmapDrawable drawable = (BitmapDrawable) ContextCompat.getDrawable(this, R.drawable.example);
-            Bitmap exampleBitmap = drawable.getBitmap();
 
-            try {
+        try {
+
+            if (!file.exists() && ContextCompat.checkSelfPermission(this, android.Manifest.permission.WRITE_EXTERNAL_STORAGE) == PackageManager.PERMISSION_GRANTED) {
+                storageDir.mkdirs();
+                BitmapDrawable drawable = (BitmapDrawable) ContextCompat.getDrawable(this, R.drawable.example);
+                Bitmap exampleBitmap = drawable.getBitmap();
+
 
                 OutputStream output = new FileOutputStream(file);
 
@@ -49,12 +48,12 @@ public class App extends Application {
 
                 Gallery.galleryAddPic(getApplicationContext(), file.getAbsolutePath());
 
-            } catch (FileNotFoundException e) {
-                e.printStackTrace();
-            } catch (IOException e) {
-                e.printStackTrace();
             }
 
+        }catch(FileNotFoundException e){
+            e.printStackTrace();
+        }catch(IOException e){
+            e.printStackTrace();
         }
     }
 }
